pyproject.toml
[tool.poetry]
name = "edgescan"
version = "2.0.0"
description = "An API client for Edgescan"
authors = ["Tyler Fisher <tylerfisher@tylerfisher.ca>"]
license = "MIT"
repository = 'https://github.com/whitfieldsdad/edgescan-client'
readme = 'README.md'

[tool.poetry.scripts]
edgescan = "edgescan.cli:cli"
script = "edgescan.cli:cli"
tool = "edgescan.cli:cli"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
inflection = "^0.5.1"
click = "^8.1.3"
requests = "^2.28.0"
urllib3 = "^1.26.9"
python-dateutil = "^2.8.2"
pyinstaller = "^5.9.0"
isort = "^5.12.0"
black = "^23.1.0"

[tool.poetry.dev-dependencies]
coverage = "^6.1.2"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
